requirements.txt
asyncpg
python-dotenv
fastmcp
markdownify
httpx
readabilipy
pydantic
